+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-*
+* Animation frame loop mode frame skipping, making loop more fluid
 
 ## [0.10.4] - 2024-07-15